+++ pkg/jwt/github/provider.go
@@ -1,67 +1,104 @@
 package github
 
 import (
-	"context"
 	"net/http"
-	"strings"
+	"sync"
 
+	"github.com/dgrijalva/jwt-go"
+	"github.com/google/go-github/github"
 	"github.com/hellofresh/janus/pkg/config"
 	"github.com/hellofresh/janus/pkg/jwt/provider"
 	"github.com/pkg/errors"
-	"golang.org/x/oauth2"
 )
 
 func init() {
-	provider.Register("github", Provider{})
+	provider.Register("github", &Provider{})
 }
 
 // Provider abstracts the authentication for github
 type Provider struct {
 	provider.Verifier
+
+	teams  []Team
+	config config.Credentials
 }
 
 // Build acts like the constructor for a provider
-func (gp Provider) Build(config config.Credentials) provider.Provider {
+func (gp *Provider) Build(config config.Credentials) provider.Provider {
 	client := NewClient()
+	teams := gp.teamConfigsToTeam(config.Github.Teams)
 
 	return &Provider{
 		Verifier: provider.NewVerifierBasket(
-			NewTeamVerifier(teamConfigsToTeam(config.Github.Teams), client),
+			NewTeamVerifier(teams, client),
 			NewOrganizationVerifier(config.Github.Organizations, client),
 		),
+		teams:  teams,
+		config: config,
 	}
 }
 
-func teamConfigsToTeam(configTeams []config.GitHubTeamConfig) []Team {
+// GetClaims returns a JWT Map Claim
+func (gp *Provider) GetClaims(httpClient *http.Client) (jwt.MapClaims, error) {
+	client := NewClient()
+
+	var (
+		wg            sync.WaitGroup
+		user          *github.User
+		usersOrgTeams OrganizationTeams
+		err           error
+	)
+
+	wg.Add(2)
+
+	go func() {
+		defer wg.Done()
+		user, err = client.CurrentUser(httpClient)
+		if err != nil {
+			err = errors.Wrap(err, "failed to get github users")
+		}
+	}()
+
+	go func() {
+		defer wg.Done()
+		usersOrgTeams, err = client.Teams(httpClient)
+		if err != nil {
+			err = errors.Wrap(err, "failed to get github teams")
+		}
+	}()
+
+	wg.Wait()
+
+	if err != nil {
+		return nil, err
+	}
+
+	return jwt.MapClaims{
+		"sub":      *user.Login,
+		"is_admin": gp.isAdmin(usersOrgTeams),
+	}, nil
+}
+
+func (gp *Provider) teamConfigsToTeam(configTeams map[string]string) []Team {
 	teams := []Team{}
-	for _, team := range configTeams {
+	for org, team := range configTeams {
 		teams = append(teams, Team{
-			Name:         team.TeamName,
-			Organization: team.OrganizationName,
+			Name:         team,
+			Organization: org,
 		})
 	}
 	return teams
 }
 
-func extractAccessToken(r *http.Request) (string, error) {
-	// We're using OAuth, start checking for access keys
-	authHeaderValue := r.Header.Get("Authorization")
-	parts := strings.Split(authHeaderValue, " ")
-	if len(parts) < 2 {
-		return "", errors.New("attempted access with malformed header, no auth header found")
+func (gp *Provider) isAdmin(usersOrgTeams OrganizationTeams) bool {
+	for _, team := range gp.teams {
+		if teams, ok := usersOrgTeams[team.Organization]; ok {
+			for _, teamUserBelongsTo := range teams {
+				if teamUserBelongsTo == gp.config.JanusAdminTeam {
+					return true
+				}
+			}
+		}
 	}
-
-	if strings.ToLower(parts[0]) != "bearer" {
-		return "", errors.New("bearer token malformed")
-	}
-
-	return parts[1], nil
-}
-
-func getClient(token string) *http.Client {
-	ctx := context.Background()
-	ts := oauth2.StaticTokenSource(
-		&oauth2.Token{AccessToken: token},
-	)
-	return oauth2.NewClient(ctx, ts)
+	return false
 }