+++ pkg/jwt/github/team_verifier.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 
 	"github.com/pkg/errors"
-	log "github.com/sirupsen/logrus"
 )
 
 // Team represents a github team within the organization
@@ -28,13 +27,8 @@ func NewTeamVerifier(teams []Team, gitHubClient Client) *TeamVerifier {
 }
 
 // Verify makes a check and return a boolean if the check was successful or not
-func (v *TeamVerifier) Verify(r *http.Request) (bool, error) {
-	accessToken, err := extractAccessToken(r)
-	if err != nil {
-		return false, errors.Wrap(err, "failed to extract access token")
-	}
-
-	usersOrgTeams, err := v.gitHubClient.Teams(getClient(accessToken))
+func (v *TeamVerifier) Verify(r *http.Request, httpClient *http.Client) (bool, error) {
+	usersOrgTeams, err := v.gitHubClient.Teams(httpClient)
 	if err != nil {
 		return false, errors.Wrap(err, "failed to get teams")
 	}
@@ -49,10 +43,5 @@ func (v *TeamVerifier) Verify(r *http.Request) (bool, error) {
 		}
 	}
 
-	log.WithFields(log.Fields{
-		"have": usersOrgTeams,
-		"want": v.teams,
-	}).Debug("not in teams")
-
-	return false, nil
+	return false, errors.New("you are not part of the allowed teams")
 }