+++ pkg/proxy/register_test.go
@@ -63,7 +63,7 @@ func TestSuccessfulProxy(t *testing.T) {
 			defer res.Body.Close()
 		}
 
-		assert.Equal(t, tc.expectedContentType, res.Header.Get("Content-Type"))
+		assert.Equal(t, tc.expectedContentType, res.Header.Get("Content-Type"), tc.description)
 		assert.Equal(t, tc.expectedCode, res.StatusCode, tc.description)
 	}
 }
@@ -71,20 +71,27 @@ func TestSuccessfulProxy(t *testing.T) {
 func createProxyDefinitions() []*Definition {
 	return []*Definition{
 		{
-			ListenPath:  "/example/*",
-			UpstreamURL: "http://www.mocky.io/v2/58c6c60710000040151b7cad",
-			Methods:     []string{"ALL"},
+			ListenPath: "/example/*",
+			Upstreams: &Upstreams{
+				Balancing: "roundrobin",
+				Targets:   []*Target{&Target{Target: "http://localhost:9089/hello-world"}},
+			},
+			Methods: []string{"ALL"},
 		},
 		{
-			ListenPath:  "/posts/*",
-			UpstreamURL: "https://jsonplaceholder.typicode.com/posts",
-			StripPath:   true,
-			Methods:     []string{"ALL"},
+			ListenPath: "/posts/*",
+			StripPath:  true,
+			Upstreams: &Upstreams{
+				Balancing: "roundrobin",
+				Targets:   []*Target{&Target{Target: "http://localhost:9089/posts"}},
+			},
+			Methods: []string{"ALL"},
 		},
 		{
 			ListenPath:  "/append/*",
-			UpstreamURL: "http://www.mocky.io/v2/58c6c60710000040151b7cad",
+			UpstreamURL: "http://localhost:9089/hello-world",
 			AppendPath:  true,
+			Upstreams:   &Upstreams{},
 			Methods:     []string{"GET"},
 		},
 	}